requirements.txt
aiogram==3.13.1
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
openai==1.43.0
upstash-redis==1.1.0
upstash-ratelimit==1.0.0

httpx<0.28

httpx<0.28
